requirements.txt
torch
transformers
sentence-transformers
xgboost
pgmpy
scikit-learn
pandas
numpy
spacy
tqdm
shap
matplotlib
ipykernel
python-dotenv
groq
tavily-python